refactor(BlackJackService): 딜러와 플레이어의 수익의 타입 변경

+++ src/main/java/dto/ParticipantWinning.java
@@ -2,7 +2,7 @@ package dto;
 
 import java.util.List;
 
-public record ParticipantWinning(int dealerProfit, List<PlayerProfit> playersWinning) {
+public record ParticipantWinning(double dealerProfit, List<PlayerProfit> playersWinning) {
     public ParticipantWinning {
         playersWinning = List.copyOf(playersWinning);
     }
